Limpieza de la solución
